+++ models/contacts.json
@@ -1,1 +1,1 @@
-[{"id":"8","name":"Anton","email":"ajrbdvrjvyd@mail.com","phone":"09811111111"},{"id":"10","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"id":"8a2a58b5-fe17-44a5-a245-651cbd6757b0","name":"Artem","email":"drkdrj@mail.com","phone":"9474794794"},{"id":"5fd39acd-303c-4f50-9d16-fdf0eeed0d30","name":"Opopopo","email":"urividrd@mail.com","phone":"4784738473874"},{"id":"6773e6d3-7608-4bb6-80d2-99d42d806628","name":"Kooooool","email":"urividrd@mail.com","phone":"555555"},{"id":"2c0a4674-f0c8-4249-bdb5-9cf994cab372","name":"car","email":"artema@mail.net","phone":"555555"},{"id":"98d2ad51-4ff9-4dfe-a435-90a968b1368f","name":"andriy","email":"artema@mail.net","phone":"555555"},{"id":"b2b3b106-89aa-4bc2-9ab6-8dd5f29b1187","name":"Artem Horiuk Leon","email":"gogoog@mail.com","phone":"09800000000"},{"id":"53275e88-7ed9-45a2-9545-96141a6b86dc","name":"Artem Horiuk Leon","phone":"09800000000"},{"id":"5bae4052-e5fb-4ca7-ae09-02b8dd8fc1c5","name":"Julia Iglesiovna","email":"ajrbdvrjvyd@mail.com","phone":"09811111111"}]
+[{"id":"8","name":"Julia Iglesiovna","phone":"09811111111"},{"id":"10","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"id":"8a2a58b5-fe17-44a5-a245-651cbd6757b0","name":"Artem","email":"drkdrj@mail.com","phone":"9474794794"},{"id":"5fd39acd-303c-4f50-9d16-fdf0eeed0d30","name":"Opopopo","email":"urividrd@mail.com","phone":"4784738473874"},{"id":"6773e6d3-7608-4bb6-80d2-99d42d806628","name":"Kooooool","email":"urividrd@mail.com","phone":"555555"},{"id":"2c0a4674-f0c8-4249-bdb5-9cf994cab372","name":"car","email":"artema@mail.net","phone":"555555"},{"id":"98d2ad51-4ff9-4dfe-a435-90a968b1368f","name":"andriy","email":"artema@mail.net","phone":"555555"},{"id":"b2b3b106-89aa-4bc2-9ab6-8dd5f29b1187","name":"Artem Horiuk Leon","email":"gogoog@mail.com","phone":"09800000000"},{"id":"53275e88-7ed9-45a2-9545-96141a6b86dc","name":"Artem Horiuk Leon","phone":"09800000000"},{"id":"5bae4052-e5fb-4ca7-ae09-02b8dd8fc1c5","name":"Julia Iglesiovna","email":"ajrbdvrjvyd@mail.com","phone":"09811111111"},{"id":"d28a29f9-f6ad-4a55-845b-084bd1694c7a","name":"Julia Iglesiovna","phone":"09811111111"},{"id":"2283900c-234a-4e73-a1c2-3f86cc768bf1","name":"Julia Iglesiovna","phone":"09811111111"}]
