+++ routes/api/contacts.js
@@ -1,4 +1,5 @@
 const express = require("express");
+const { scheme } = require('../../middlewares/validations')
 const {
   listContacts,
   getContactById,
@@ -6,26 +7,17 @@ const {
   addContact,
   updateContact,
 } = require("../../models/contacts");
-const Joi = require("joi");
+
 const router = express.Router();
 
-const scheme = Joi.object({
-  name: Joi.string().required().min(2).max(30),
-  phone: Joi.string().required(),
-  email: Joi.string()
-    .required()
-    .email({
-      minDomainSegments: 2,
-      tlds: { allow: ["com", "net"] },
-    }),
-});
+
 
 router.get("/", async (req, res, next) => {
   try {
     const contacts = await listContacts();
-    res.json({
+    res.json(
       contacts
-    });
+    );
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
@@ -90,13 +82,14 @@ router.put("/:contactId", async (req, res, next) => {
         message: error.message,
       });
     }
-
     const changedContact = await updateContact(contactId, req.body);
+
     if (!changedContact) {
       return res.status(404).json({
         message: error.message,
       });
     }
+     
     res.status(200).json({
       changedContact
     });
